Fixed progress callback for curl (incorrect parameter declaration).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7216 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/addons/network_http.hpp
@@ -76,9 +76,8 @@ private:
     bool          downloadFileInternal(const std::string &file,
                                        const std::string &save_filename,
                                        bool is_asynchron);
-    static int    progressDownload(void *clientp, float dltotal, float dlnow,
-                                   float ultotal, float ulnow);
-
+    static int    progressDownload(void *clientp, double dltotal, double dlnow,
+                                   double ultotal, double ulnow);
 public:
                   NetworkHttp();
                  ~NetworkHttp();
